% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendroSense-package.R
\docType{package}
\name{dendroSense-package}
\alias{dendroSense}
\alias{dendroSense-package}
\title{dendroSense: Analyzing dendrometers data the easy way.}
\description{
DendroSense is an R package and Shiny app for analyzing dendrometers data.
}
\author{
\strong{Maintainer}: Hassan (Roland) Nasser \email{hassan.nasser@me.com} (\href{https://orcid.org/0000-0003-1821-3234}{ORCID})

}
\keyword{internal}
